[refactor][style] 스타일 수정
