+++ src/app/blog/(posts)/post-9/page.mdx
@@ -3,7 +3,7 @@ export const metadata = {
   description: "TypeScript를 처음 시작하는 사람들을 위한 가이드입니다.",
   publishDate: "2025-03-09T00:00:00Z",
   thumbnailUrl: "/assets/blog/thumbnail-ts.png",
-  tags: [],
+  tags: ["TypeScript"],
 };
 
 <BlogHeader props={{ ...metadata }} />
