--- src/components/blog/BlogWritter.tsx
@@ -1,15 +1,0 @@
-function BlogWritter() {
-  return (
-    <div className="inline-flex items-center gap-0.5">
-      <img
-        id="profile-avatar"
-        src="/assets/profile-avatar.webp"
-        alt="profile-avatar"
-        className="h-5 w-5 rounded-full"
-      />
-      <label htmlFor="profile-avatar">dev-oh</label>
-    </div>
-  );
-}
-
-export default BlogWritter;
