+++ src/components/common/LeftSidebar.tsx
@@ -23,7 +23,7 @@ const LeftSidebar: React.FC = () => {
 
   return (
     <aside
-      className={`fixed z-10 h-screen space-y-2 rounded-r-3xl border-r border-gray-300 bg-gray-300/50 py-12 shadow-lg ${isOpen || isPinned ? "w-60 min-w-60" : "w-18 min-w-18 opacity-0"}`}
+      className={`bg-background/50 fixed top-18 z-10 h-[calc(100vh-4.5rem)] space-y-2 border-r border-gray-300 pt-18 shadow-lg ${isOpen || isPinned ? "w-60 min-w-60" : "w-18 min-w-18 opacity-0"}`}
       onMouseEnter={() => !isPinned && setIsOpen(true)}
       onMouseLeave={() => !isPinned && setIsOpen(false)}
     >
